refactor: ReservationRepository를 인터페이스에서 추상메서드로 변경
- 구현체들에서 공통으로 사용되는 상수들의 접근제어자를 protected로 변경하여 외부에서 접근 어렵도록 변경
- 패키지 구조 변경
