+++ README.md
@@ -29,7 +29,8 @@
   - [ ] 비즈니스 로직을 담당하는 객체를 별도로 만들어 사용한다.
 
 # 리뷰 후 수정 필요한 사항
-- [ ] 인터페이스 내 변수들 제거 및 다른 방식으로 중복 제거
+- [x] 인터페이스 내 변수들 제거 및 다른 방식으로 중복 제거
+  - 인터페이스 대신 추상클래스 사용하는 방식으로 해결
 - [x] 기존 jdbc 관련 config를 application.yml을 이용하는 방향으로 변경
 - [ ] ReservationException 정의하는 부분 super 추가
 - [ ] 에러 발생시 적절한 HTTP 상태 코드 응답하도록 변경
