+++ src/test/java/nextstep/repository/reservation/ConsoleReservationRepositoryTest.java
@@ -1,4 +1,4 @@
-package nextstep.repository;
+package nextstep.repository.reservation;
 
 import nextstep.domain.Reservation;
 import nextstep.domain.Theme;
